refactor(api): hoist createE2EGatedSingleton to shared/api; JSDoc/comment polish (PR #649 suggestions)
