+++ src/shared/api/createE2EGatedSingleton.ts
