+++ src/shared/api/fmp/getCongressTradesProvider.ts
@@ -1,4 +1,4 @@
-import { createE2EGatedSingleton } from './createE2EGatedSingleton';
+import { createE2EGatedSingleton } from '@/shared/api/createE2EGatedSingleton';
 import type { CongressTradesProvider } from '@y0ngha/siglens-core';
 import { CachedCongressTradesProvider } from './CachedCongressTradesProvider';
 import { FmpCongressTradesClient } from './congressTradesClient';
