+++ src/shared/api/fmp/symKey.ts
@@ -1,8 +1,6 @@
 /**
  * 심볼을 대문자로 정규화한다.
  *
- * 캐시 키에서 대소문자 차이로 엔트리가 분리되는 것을 막기 위해
- * CachedFundamentalProvider, CachedFinancialStatementsProvider,
- * CachedCongressTradesProvider 세 곳이 동일하게 사용한다.
+ * 캐시 키 정규화 — 대소문자 차이로 인해 Redis 엔트리가 중복 생성되는 것을 방지한다.
  */
 export const sym = (s: string): string => s.toUpperCase();
